backend/requirements.txt
fastapi
uvicorn[standard]
pandas
scikit-learn
pydantic
numpy
google-generativeai
python-multipart
python-dotenv